requirements.txt
Flask==2.3.2
yt-dlp
python-dotenv==1.0.0
redis>=4.5.0
rq==1.11.1
rq-dashboard>=0.8.6
gunicorn==21.2.0
